chore: re-license `cosmic-manager` CLI
Re-licesing it because the code in src/config.rs was taken from cosmic-ctl,
which is also a project that I made, but I need to comply with its license.
So instead of re-licensing just one file, I decided to re-license the entire
CLI, because I think GPL-3.0 fits it better anyway.